+++ Utilities/DebugConsole/Api.cs
@@ -171,28 +171,32 @@ namespace Sheepy.PhoenixPt.DebugConsole {
       #endregion
 
       #region Gui Tree
+      internal static object GuiTreeApi ( object root ) => DumpGui( null, root ?? DefaultDumpSubject() );
+
       [ ConsoleCommand( Command = "dump_gui", Description = " - Dump module component tree." ) ]
-      public static void ConsoleCommandGuiDump ( IConsole console ) { try {
-         var level = GameUtl.CurrentLevel();
-         if ( level == null || level.CurrentState != Level.State.Playing ) {
-            DumpGui( console, level.gameObject );
-            return;
+      public static void ConsoleCommandGuiDump ( IConsole console, string[] names ) { try {
+         if ( names?.Length > 0 ) {
+            foreach ( var name in names ) {
+               var obj = GameObject.Find( name.Trim() );
+               if ( obj == null ) console.WriteLine( "Cannot find " + name );
+               else DumpGui( console, obj );
+            }
          }
-         if ( level.GetComponent<MenuLevelController>() != null ) {
-            var view = level.GetComponent<HomeScreenView>();
-            DumpGui( console, view.HomeScreenModules );
-         }
-         else if ( level.GetComponent<GeoLevelController>() != null ) {
-            var view = level.GetComponent<GeoscapeView>();
-            DumpGui( console, view.GeoscapeModules );
-         }
-         else if ( level.GetComponent<TacticalLevelController>() != null ) {
-            var view = level.GetComponent<TacticalView>();
-            DumpGui( console, view.TacticalModules );
-         }
+         DumpGui( console, DefaultDumpSubject() );
       } catch ( Exception ex ) { Error( ex ); } }
 
-      internal static object GuiTreeApi ( object root ) => DumpGui( null, root );
+      private static object DefaultDumpSubject () {
+         var level = GameUtl.CurrentLevel();
+         if ( level == null || level.CurrentState != Level.State.Playing )
+            return level.gameObject;
+         if ( level.GetComponent<MenuLevelController>() != null )
+            return level.GetComponent<HomeScreenView>()?.HomeScreenModules;
+         if ( level.GetComponent<GeoLevelController>() != null )
+            return level.GetComponent<GeoscapeView>()?.GeoscapeModules;
+         if ( level.GetComponent<TacticalLevelController>() != null )
+            return level.GetComponent<TacticalView>()?.TacticalModules;
+         return null;
+      }
 
       internal static object DumpGui ( IConsole output, object root ) {
          if ( root is Component c ) root = c.gameObject;
@@ -209,15 +213,12 @@ namespace Sheepy.PhoenixPt.DebugConsole {
          if ( logged.Contains( e ) ) return;
          logged.Add( e );
          var rect = e.GetComponent<Transform>();
-         Output( output, "{0}> {1} {2}{3} Layer {4} / Pos {5} Scale {6} Rot {7}", prefix, e.name, TypeName( e ), e.activeSelf ? "" : " (Inactive)", e.layer,
-            rect?.localPosition, rect?.localScale, rect?.localRotation );
+         Output( output, "{0}> '{1}' {2}{3} Layer {4} : {5}", prefix, e.name, TypeName( e ), e.activeSelf ? "" : " (Inactive)", e.layer, ToString( rect ) );
          if ( output == null || prefix.Length == 0 )
             foreach ( var c in e.GetComponents<Component>() ) {
                var typeName = TypeName( c );
-               if ( c is Transform cRect ) {
-                  if ( c != rect )
-                     Output( output, "{0}...{1} Pos {2} Scale {3} Rotate {4}", prefix, typeName, cRect.localPosition, cRect.localScale, cRect.localRotation );
-               } else if ( c is UnityEngine.UI.Text txt )
+               if ( c is Transform cRect ) ;
+               else if ( c is UnityEngine.UI.Text txt )
                   Output( output, "{0}...{1} {2} {3}", prefix, typeName, txt.color, txt.text );
                else if ( c is I2.Loc.Localize loc )
                   Output( output, "{0}...{1} {2}", prefix, typeName, loc.mTerm );
@@ -232,9 +233,21 @@ namespace Sheepy.PhoenixPt.DebugConsole {
             DumpComponents( output, prefix + "  ", logged, e.transform.GetChild( i ).gameObject );
       }
 
+      private static Func<string> ToString ( Transform t ) { return () => {
+         if ( t == null ) return "null";
+         var result = string.Format( "Pos {0} Scale {1} Rotate {2}", t.localPosition, t.localScale, t.localRotation );
+         return result.Replace( ".0,", "," ).Replace( ".0)", ")" )
+            .Replace( "Pos (0, 0, 0)", "" )
+            .Replace( "Scale (1, 1, 1)", "" )
+            .Replace( "Rotate (0, 0, 0, 1)", "" )
+            .Trim();
+      }; }
+
       private static void Output ( IConsole output, object param, params object[] args ) {
-         if ( output == null ) Info( param, args );
-         else output.WriteLine( param?.ToString(), args );
+         if ( output == null ) { Info( param, args ); return; }
+         for ( int i = 0, len = args.Length ; i < len ; i++ )
+            if ( args[i] is Func<string> eval ) args[i] = eval();
+         output.WriteLine( param?.ToString(), args );
       }
 
       private static string TypeName ( object e ) => e?.GetType().FullName.Replace( "UnityEngine.", "UE." );